backend/requirements.txt
# BiPay Backend Requirements
fastapi==0.104.1
uvicorn[standard]==0.24.0
pymongo==4.5.0
motor==3.3.1
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
opencv-python==4.8.1.78
numpy==1.24.4
scikit-learn==1.3.2
pandas==2.1.4
loguru==0.7.2
python-multipart==0.0.6
email-validator==2.1.0
python-dateutil==2.8.2
